Change Makefile to use mpif90
changed from specifying the explicit path, which becomes stale
after a while. The new change requires the user to have a proper
fortran + mpi compiler set up. As of this writing, that is achieved
on tyndall's rocks cluster by adding "module add pgi/12.9 pgi-mpi/12.9"
to the end of your shell's .rc file.